refactor: improved quality of database interaction
