+++ package.json
@@ -21,19 +21,13 @@
     "lint:fix": "eslint . --fix",
     "prettier:fix": "npx prettier --write .",
     "prepare": "husky install",
-    "typeorm-postgres": "npm run build && npx typeorm -d dist/config/databases/postgres/postgres.source.js",
+    "typeorm": "npm run build && npx typeorm -d dist/databases/data-source.js",
     "migration:create": "npx typeorm migration:create",
-    "migration:generate-postgres": "npm run typeorm-postgres -- migration:generate",
-    "migration:run-postgres": "npm run typeorm-postgres -- migration:run",
-    "migration:revert-postgres": "npm run typeorm-postgres -- migration:revert",
-    "migration:show-postgres": "npm run typeorm-postgres -- migration:show",
-    "migration:pending-postgres": "npm run typeorm-postgres migration:run",
-    "typeorm-mongodb": "npm run build && npx typeorm -d dist/config/databases/mongodb/mongodb.source.js",
-    "migration:generate-mongodb": "npm run typeorm-mongodb -- migration:generate",
-    "migration:run-mongodb": "npm run typeorm-mongodb -- migration:run",
-    "migration:revert-mongodb": "npm run typeorm-mongodb -- migration:revert",
-    "migration:show-mongodb": "npm run typeorm-mongodb -- migration:show",
-    "migration:pending-mongodb": "npm run typeorm-mongodb migration:run"
+    "migration:generate": "npm run typeorm -- migration:generate",
+    "migration:run": "npm run typeorm -- migration:run",
+    "migration:revert": "npm run typeorm -- migration:revert",
+    "migration:show": "npm run typeorm -- migration:show",
+    "migration:pending": "npm run typeorm migration:run"
   },
   "dependencies": {
     "@nestjs/cache-manager": "^2.2.2",
